AG-16578 Hoist anyOverlap and grid cell-sizing into ag-charts-core
